requirements.txt
# These requirements are only necessary when developing on 

# development
Werkzeug>=1.0,<2.1
django-debug-toolbar>=2.2,<3.6
django-extensions>=2.2,<3.3
psycopg2-binary==2.9.6

# Sandbox
Whoosh>=2.7,<2.8
django-environ>=0.4,<0.10
django-redis>=4.12,<5.3
pysolr>=3.9,<3.10
redis>=3.5,<4.4
requests>=2.25,<3
uWSGI>=2.0.19,<2.1
whitenoise>=5.2,<6.3

# Linting
flake8==4.0.1
flake8-debugger==4.1.2
isort==5.10.1

# Helpers
pyprof2calltree>=1.4,<1.5
ipdb>=0.13,<0.14
ipython>=7.12,<9

# Country data
pycountry

django-widget-tweaks==1.4.12
django-haystack==3.2.1
# django-haystack[elasticsearch]==7.17.9
django-treebeard==4.7
sorl-thumbnail==12.9.0
# easy-thumbnails==2.8.5
django_tables2==2.5.3
django_phonenumbers==1.0.1
django_phonenumber_field==7.0.2
Babel==2.12.1
django_extra_views==0.14.0
purl==1.6

Pillow==9.5.0
# mysqlclient==2.1.1
python-decouple==3.8